// src/types/auth.ts
import { ApiResponse } from "./api-response.js";

/** Supported OAuth grant types. */
// export type AuthGrantType =
//   | "client_credentials"
//   | "refresh_token";

/** OAuth token details returned by Nomba. */
export interface TokenData {
  /** The accountId of the merchant/business connecting to the Nomba system via API. */
  businessId: string;

  /**
   * A JWT (JSON Web Token) used to grant access to resources
   * on the Nomba API.
   *
   * Required string length: 300 - 600.
   */
  access_token: string;

  /**
   * A Base64 encoded string used to refresh an expired access_token.
   *
   * Required string length: 30 - 60.
   */
  refresh_token: string;

  /**
   * The date and time this entity was created.
   *
   * This value uses Coordinated Universal Time (UTC) and ISO 8601 format:
   * YYYY-MM-DDThh:mm:ssZ.
   */
  expiresAt: string;
}

/** Credentials used to obtain an access token. */
export interface IssueAccessTokenRequestBody {
  /** The grant type of the authentication. */
  grant_type: "client_credentials";

  // /**
  //  * The client's ID as obtained from your Nomba dashboard.
  //  *
  //  * Required string length: 36.
  //  */
  // client_id: string;

  // /**
  //  * The client's secret as obtained from your Nomba dashboard.
  //  * It is securely generated by the Nomba system.
  //  *
  //  * Required string length: 44.
  //  */
  // client_secret: string;
}

/** Response returned when an access token is issued. */
export type IssueAccessTokenResponse =
  ApiResponse<TokenData>;

/** Refresh token used to obtain a new access token. */
export interface RefreshAccessTokenRequestBody {
  /** The grant type of the authentication. */
  grant_type: "refresh_token";

  /**
   * A Base64 encoded string used to refresh an expired access_token.
   *
   * Required string length: 30 - 60.
   */
  refresh_token: string;
}

/** Response returned when an access token is refreshed. */
export type RefreshAccessTokenResponse =
  ApiResponse<TokenData>;

/** Credentials used to revoke an access token. */
export interface RevokeAccessTokenRequestBody {
  // /**
  //  * The client's ID as obtained from your Nomba dashboard.
  //  *
  //  * Required string length: 36.
  //  */
  // clientId: string;

  /**
   * The JWT (JSON Web Token) token to be revoked.
   *
   * Required string length: 300 - 600.
   */
  // access_token: string;
}

/** Response returned after revoking an access token. */
export interface RevokeAccessTokenResponse {
  /** Tells if the access_token has been revoked. */
  code: string;

  /** Describes the result of the action performed. */
  description: string;

  data?: unknown;
}